chore: Update shape_estimation library and node names
refactor: relocate headers

refactor: add namespace

refactor: update shape_estimation  namespace on the user side
