+++ perception/shape_estimation/include/autoware/shape_estimation/corrector/bus_corrector.hpp
@@ -12,28 +12,34 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SHAPE_ESTIMATION__CORRECTOR__CAR_CORRECTOR_HPP_
-#define SHAPE_ESTIMATION__CORRECTOR__CAR_CORRECTOR_HPP_
+#ifndef AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__BUS_CORRECTOR_HPP_
+#define AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__BUS_CORRECTOR_HPP_
 
-#include "shape_estimation/corrector/vehicle_corrector.hpp"
+#include "autoware/shape_estimation/corrector/vehicle_corrector.hpp"
 #include "utils.hpp"
-
-class CarCorrector : public VehicleCorrector
+namespace autoware::shape_estimation
+{
+namespace corrector
+{
+class BusCorrector : public VehicleCorrector
 {
 public:
-  explicit CarCorrector(const bool use_reference_yaw = false) : VehicleCorrector(use_reference_yaw)
+  explicit BusCorrector(const bool use_reference_yaw = false) : VehicleCorrector(use_reference_yaw)
   {
     corrector_utils::CorrectionBBParameters params;
-    params.min_width = 1.2;
-    params.max_width = 2.5;
+    params.min_width = 2.0;
+    params.max_width = 3.2;
     params.default_width = (params.min_width + params.max_width) * 0.5;
-    params.min_length = 3.0;
-    params.max_length = 5.8;
-    params.default_length = (params.min_length + params.max_length) * 0.5;
+    params.min_length = 5.0;
+    params.max_length = 17.0;
+    params.default_length = 7.0;
     setParams(params);
   }
 
-  ~CarCorrector() = default;
+  ~BusCorrector() = default;
 };
 
-#endif  // SHAPE_ESTIMATION__CORRECTOR__CAR_CORRECTOR_HPP_
+}  // namespace corrector
+}  // namespace autoware::shape_estimation
+
+#endif  // AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__BUS_CORRECTOR_HPP_
