+++ perception/shape_estimation/include/autoware/shape_estimation/corrector/corrector.hpp
@@ -12,4 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#ifndef AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__CORRECTOR_HPP_
+#define AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__CORRECTOR_HPP_
+
+#include "autoware/shape_estimation/corrector/bicycle_corrector.hpp"
+#include "autoware/shape_estimation/corrector/bus_corrector.hpp"
+#include "autoware/shape_estimation/corrector/car_corrector.hpp"
+#include "autoware/shape_estimation/corrector/corrector_interface.hpp"
 #include "autoware/shape_estimation/corrector/no_corrector.hpp"
+#include "autoware/shape_estimation/corrector/reference_shape_size_corrector.hpp"
+#include "autoware/shape_estimation/corrector/trailer_corrector.hpp"
+#include "autoware/shape_estimation/corrector/truck_corrector.hpp"
+
+#endif  // AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__CORRECTOR_HPP_
