+++ perception/shape_estimation/include/autoware/shape_estimation/corrector/trailer_corrector.hpp
@@ -12,11 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SHAPE_ESTIMATION__CORRECTOR__TRAILER_CORRECTOR_HPP_
-#define SHAPE_ESTIMATION__CORRECTOR__TRAILER_CORRECTOR_HPP_
+#ifndef AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__TRAILER_CORRECTOR_HPP_
+#define AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__TRAILER_CORRECTOR_HPP_
 
-#include "shape_estimation/corrector/vehicle_corrector.hpp"
+#include "autoware/shape_estimation/corrector/vehicle_corrector.hpp"
 #include "utils.hpp"
+
+namespace autoware::shape_estimation
+{
+namespace corrector
+{
 
 // Generally speaking, trailer would be much larger than bus and truck.
 // But currently we do not make large differences among bus/truck/trailer
@@ -40,4 +45,7 @@ public:
   ~TrailerCorrector() = default;
 };
 
-#endif  // SHAPE_ESTIMATION__CORRECTOR__TRAILER_CORRECTOR_HPP_
+}  // namespace corrector
+}  // namespace autoware::shape_estimation
+
+#endif  // AUTOWARE__SHAPE_ESTIMATION__CORRECTOR__TRAILER_CORRECTOR_HPP_
