+++ perception/shape_estimation/include/autoware/shape_estimation/filter/filter_interface.hpp
@@ -12,22 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/shape_estimation/filter/trailer_filter.hpp"
+#ifndef AUTOWARE__SHAPE_ESTIMATION__FILTER__FILTER_INTERFACE_HPP_
+#define AUTOWARE__SHAPE_ESTIMATION__FILTER__FILTER_INTERFACE_HPP_
+
+#include <autoware_perception_msgs/msg/shape.hpp>
+#include <geometry_msgs/msg/pose.hpp>
+
+#include <string>
 
 namespace autoware::shape_estimation
 {
 namespace filter
 {
 
-bool TrailerFilter::filter(
-  const autoware_perception_msgs::msg::Shape & shape,
-  [[maybe_unused]] const geometry_msgs::msg::Pose & pose)
+class ShapeEstimationFilterInterface
 {
-  constexpr float min_width = 1.5;
-  constexpr float max_width = 3.2;
-  constexpr float max_length = 25.0;  // maximum Full-TRAILER size in JAPAN(normally upto 18m)
-  return utils::filterVehicleBoundingBox(shape, min_width, max_width, max_length);
-}
+public:
+  ShapeEstimationFilterInterface() {}
+
+  virtual ~ShapeEstimationFilterInterface() {}
+
+  virtual bool filter(
+    const autoware_perception_msgs::msg::Shape & shape, const geometry_msgs::msg::Pose & pose) = 0;
+};
 
 }  // namespace filter
 }  // namespace autoware::shape_estimation
+
+#endif  // AUTOWARE__SHAPE_ESTIMATION__FILTER__FILTER_INTERFACE_HPP_
