+++ perception/shape_estimation/include/autoware/shape_estimation/filter/truck_filter.hpp
@@ -12,21 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SHAPE_ESTIMATION__FILTER__NO_FILTER_HPP_
-#define SHAPE_ESTIMATION__FILTER__NO_FILTER_HPP_
+#ifndef AUTOWARE__SHAPE_ESTIMATION__FILTER__TRUCK_FILTER_HPP_
+#define AUTOWARE__SHAPE_ESTIMATION__FILTER__TRUCK_FILTER_HPP_
 
-#include "shape_estimation/filter/filter_interface.hpp"
-
-class NoFilter : public ShapeEstimationFilterInterface
+#include "autoware/shape_estimation/filter/filter_interface.hpp"
+#include "utils.hpp"
+namespace autoware::shape_estimation
+{
+namespace filter
+{
+class TruckFilter : public ShapeEstimationFilterInterface
 {
 public:
-  NoFilter() = default;
+  TruckFilter() = default;
 
-  ~NoFilter() = default;
+  ~TruckFilter() = default;
 
   bool filter(
     const autoware_perception_msgs::msg::Shape & shape,
     const geometry_msgs::msg::Pose & pose) override;
 };
+}  // namespace filter
+}  // namespace autoware::shape_estimation
 
-#endif  // SHAPE_ESTIMATION__FILTER__NO_FILTER_HPP_
+#endif  // AUTOWARE__SHAPE_ESTIMATION__FILTER__TRUCK_FILTER_HPP_
