+++ perception/shape_estimation/include/autoware/shape_estimation/model/cylinder.hpp
@@ -12,10 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SHAPE_ESTIMATION__MODEL__CYLINDER_HPP_
-#define SHAPE_ESTIMATION__MODEL__CYLINDER_HPP_
+#ifndef AUTOWARE__SHAPE_ESTIMATION__MODEL__CYLINDER_HPP_
+#define AUTOWARE__SHAPE_ESTIMATION__MODEL__CYLINDER_HPP_
 
-#include "shape_estimation/model/model_interface.hpp"
+#include "autoware/shape_estimation/model/model_interface.hpp"
+
+namespace autoware::shape_estimation
+{
+namespace model
+{
 
 class CylinderShapeModel : public ShapeEstimationModelInterface
 {
@@ -33,4 +38,7 @@ public:
     geometry_msgs::msg::Pose & pose_output) override;
 };
 
-#endif  // SHAPE_ESTIMATION__MODEL__CYLINDER_HPP_
+}  // namespace model
+}  // namespace autoware::shape_estimation
+
+#endif  // AUTOWARE__SHAPE_ESTIMATION__MODEL__CYLINDER_HPP_
