--- perception/shape_estimation/include/shape_estimation/corrector/corrector.hpp
@@ -1,27 +1,0 @@
-// Copyright 2018 Autoware Foundation. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef SHAPE_ESTIMATION__CORRECTOR__CORRECTOR_HPP_
-#define SHAPE_ESTIMATION__CORRECTOR__CORRECTOR_HPP_
-
-#include "shape_estimation/corrector/bicycle_corrector.hpp"
-#include "shape_estimation/corrector/bus_corrector.hpp"
-#include "shape_estimation/corrector/car_corrector.hpp"
-#include "shape_estimation/corrector/corrector_interface.hpp"
-#include "shape_estimation/corrector/no_corrector.hpp"
-#include "shape_estimation/corrector/reference_shape_size_corrector.hpp"
-#include "shape_estimation/corrector/trailer_corrector.hpp"
-#include "shape_estimation/corrector/truck_corrector.hpp"
-
-#endif  // SHAPE_ESTIMATION__CORRECTOR__CORRECTOR_HPP_
