--- perception/shape_estimation/include/shape_estimation/filter/filter.hpp
@@ -1,25 +1,0 @@
-// Copyright 2018 Autoware Foundation. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef SHAPE_ESTIMATION__FILTER__FILTER_HPP_
-#define SHAPE_ESTIMATION__FILTER__FILTER_HPP_
-
-#include "shape_estimation/filter/bus_filter.hpp"
-#include "shape_estimation/filter/car_filter.hpp"
-#include "shape_estimation/filter/filter_interface.hpp"
-#include "shape_estimation/filter/no_filter.hpp"
-#include "shape_estimation/filter/trailer_filter.hpp"
-#include "shape_estimation/filter/truck_filter.hpp"
-
-#endif  // SHAPE_ESTIMATION__FILTER__FILTER_HPP_
